+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 # v0.17.0
+- Added support to a new attribute type in the entities called `coordinates` to store geolocation data with latitude and longitude values.
 - Update AGP version to 8.13.2
 
 # v0.16.4